Remove delay_timer and implement Delay for Timer
